refactor: Rest controller 이름을 일반화
