+++ src/main/java/shopping/controller/view/AuthViewController.java
@@ -1,0 +1,13 @@
+package shopping.controller.view;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AuthViewController {
+
+    @GetMapping("/login")
+    public String getLoginPage() {
+        return "login";
+    }
+}
