+++ src/main/java/shopping/controller/view/CartViewController.java
@@ -1,0 +1,13 @@
+package shopping.controller.view;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class CartViewController {
+
+    @GetMapping("/cart")
+    public String getAllCartProductsPage() {
+        return "cart";
+    }
+}
