+++ src/main/java/shopping/controller/view/MainViewController.java
@@ -1,4 +1,4 @@
-package shopping.controller;
+package shopping.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -6,11 +6,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import shopping.service.ProductService;
 
 @Controller
-public class MainController {
+public class MainViewController {
 
     private final ProductService productService;
 
-    public MainController(final ProductService productService) {
+    public MainViewController(final ProductService productService) {
         this.productService = productService;
     }
